- cleaned up data_manipulation.py and the corresponding tests
